refactor: convert settings dropdown to dialog
squashed
