+++ src/components/transport.tsx
@@ -1,32 +1,18 @@
-import { useMutation } from "@tanstack/react-query";
 import {
   CheckIcon,
   ChevronsUpDownIcon,
   CircleHelpIcon,
-  ClipboardIcon,
-  DownloadIcon,
-  FolderIcon,
   PauseIcon,
   PlayIcon,
   SettingsIcon,
   SlidersHorizontalIcon,
-  Trash2Icon,
-  UploadIcon,
 } from "lucide-react";
 import { useRef, useState } from "react";
-import { toast } from "sonner";
 import * as Tone from "tone";
 import { useDraftInput } from "../hooks/use-draft-input";
 import { useTransport } from "../hooks/use-transport";
 import { useWindowEvent } from "../hooks/use-window-event";
-import {
-  copyABCToClipboard,
-  downloadABCFile,
-  exportABC,
-} from "../lib/abc-export";
-import { deleteAsset, saveAsset } from "../lib/asset-store";
-import { audioManager, GM_PROGRAMS, loadAudioFile } from "../lib/audio";
-import { downloadMidiFile, exportMidi } from "../lib/midi-export";
+import { audioManager, GM_PROGRAMS } from "../lib/audio";
 import { useProjectStore } from "../stores/project-store";
 import { COMMON_TIME_SIGNATURES, type GridSnap } from "../types";
 import { Button } from "./ui/button";
@@ -40,13 +26,9 @@ import {
 } from "./ui/command";
 import {
   DropdownMenu,
-  DropdownMenuCheckboxItem,
   DropdownMenuContent,
-  DropdownMenuItem,
   DropdownMenuRadioGroup,
   DropdownMenuRadioItem,
-  DropdownMenuSeparator,
-  DropdownMenuShortcut,
   DropdownMenuTrigger,
 } from "./ui/dropdown-menu";
 import { Popover, PopoverContent, PopoverTrigger } from "./ui/popover";
@@ -227,31 +209,25 @@ function InstrumentCombobox({
 }
 
 type TransportProps = {
-  onProjectSettingsClick: () => void;
+  onSettingsClick: () => void;
   onHelpClick: () => void;
-  onProjectsClick: () => void;
   onMixerClick: () => void;
 };
 
 export function Transport({
-  onProjectSettingsClick,
+  onSettingsClick,
   onHelpClick,
-  onProjectsClick,
   onMixerClick,
 }: TransportProps) {
   const {
-    audioFileName,
-    audioAssetKey,
     tempo,
     timeSignature,
-    notes,
     midiProgram,
     midiMuted,
     audioMuted,
     metronomeEnabled,
     autoScrollEnabled,
     gridSnap,
-    showDebug,
     setTempo,
     setTimeSignature,
     setMidiProgram,
@@ -260,14 +236,8 @@ export function Transport({
     setMetronomeEnabled,
     setAutoScrollEnabled,
     setGridSnap,
-    setShowDebug,
-    setAudioFile,
-    setAudioOffset,
-    setAudioView,
-    clearAudioFile,
   } = useProjectStore();
 
-  const fileInputRef = useRef<HTMLInputElement>(null);
   const tapTimesRef = useRef<number[]>([]);
 
   const tempoInput = useDraftInput({
@@ -276,46 +246,6 @@ export function Transport({
     min: 30,
     max: 300,
   });
-
-  const loadAudioMutation = useMutation({
-    mutationFn: async (file: File) => {
-      const { buffer, audioView } = await loadAudioFile(file);
-
-      // Save audio to IndexedDB for persistence
-      const assetKey = await saveAsset(file);
-      setAudioFile(file.name, buffer.duration, assetKey);
-
-      audioManager.player.buffer = buffer;
-      audioManager.player.sync().start(0);
-      setAudioOffset(0);
-
-      setAudioView(audioView);
-    },
-  });
-
-  const handleLoadClick = () => {
-    fileInputRef.current?.click();
-  };
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (file) {
-      loadAudioMutation.mutate(file);
-    }
-    // Reset input so same file can be selected again
-    e.target.value = "";
-  };
-
-  const handleRemoveAudio = async () => {
-    // Delete from IndexedDB if we have a key
-    if (audioAssetKey) {
-      await deleteAsset(audioAssetKey);
-    }
-    // Clear the audio buffer in the player
-    audioManager.clearAudioBuffer();
-    // Clear store state
-    clearAudioFile();
-  };
 
   // Keyboard shortcuts: M=metronome, Ctrl+F=auto-scroll, Shift+1/2=mute (Space is handled by PlayPauseButton)
   useWindowEvent("keydown", (e) => {
@@ -396,80 +326,11 @@ export function Transport({
     }
   };
 
-  const handleExportMidi = () => {
-    const midiData = exportMidi({
-      notes,
-      tempo,
-      timeSignature,
-      trackName: audioFileName
-        ? audioFileName.replace(/\.[^.]+$/, "")
-        : "Piano Roll",
-    });
-
-    // Generate filename with timestamp
-    const now = new Date();
-    const timestamp = now
-      .toISOString()
-      .replace(/[T:]/g, "-")
-      .replace(/\.\d+Z$/, "");
-    const fileName = `toy-midi-export-${timestamp}.mid`;
-
-    downloadMidiFile(midiData, fileName);
-  };
-
-  const handleExportABC = () => {
-    const abcText = exportABC({
-      notes,
-      tempo,
-      timeSignature,
-      title: audioFileName ? audioFileName.replace(/\.[^.]+$/, "") : "Untitled",
-    });
-
-    // Generate filename with timestamp
-    const now = new Date();
-    const timestamp = now
-      .toISOString()
-      .replace(/[T:]/g, "-")
-      .replace(/\.\d+Z$/, "");
-    const fileName = `toy-midi-export-${timestamp}.abc`;
-
-    downloadABCFile(abcText, fileName);
-  };
-
-  const handleCopyABCToClipboard = async () => {
-    try {
-      const abcText = exportABC({
-        notes,
-        tempo,
-        timeSignature,
-        title: audioFileName
-          ? audioFileName.replace(/\.[^.]+$/, "")
-          : "Untitled",
-      });
-
-      await copyABCToClipboard(abcText);
-      toast.success("ABC notation copied to clipboard");
-    } catch (error) {
-      console.error("Failed to copy ABC notation:", error);
-      toast.error("Failed to copy to clipboard");
-    }
-  };
-
   return (
     <div
       data-testid="transport"
       className="flex items-center gap-2 px-3 py-2 bg-neutral-800 border-b border-neutral-700"
     >
-      {/* Hidden file input */}
-      <input
-        data-testid="audio-file-input"
-        ref={fileInputRef}
-        type="file"
-        accept="audio/*"
-        onChange={handleFileChange}
-        className="hidden"
-      />
-
       {/* Play/Pause button */}
       <PlayPauseButton />
 
@@ -587,114 +448,16 @@ export function Transport({
       {/* Spacer */}
       <div className="flex-1" />
 
-      {/* Settings dropdown */}
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <Button
-            data-testid="settings-button"
-            variant="ghost"
-            size="icon"
-            title="Settings"
-          >
-            <SettingsIcon className="size-5" />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent align="end" className="w-48">
-          {/* Project Settings */}
-          <DropdownMenuItem
-            data-testid="project-settings-button"
-            onClick={onProjectSettingsClick}
-          >
-            <SettingsIcon className="size-4" />
-            Project Settings
-          </DropdownMenuItem>
-
-          {/* Projects */}
-          <DropdownMenuItem
-            data-testid="projects-button"
-            onClick={onProjectsClick}
-          >
-            <FolderIcon className="size-4" />
-            Projects
-          </DropdownMenuItem>
-
-          <DropdownMenuSeparator />
-
-          {/* Load Audio */}
-          <DropdownMenuItem
-            data-testid="load-audio-button"
-            onClick={handleLoadClick}
-            disabled={loadAudioMutation.isPending}
-          >
-            <UploadIcon className="size-4" />
-            {loadAudioMutation.isPending ? "Loading..." : "Load Audio"}
-          </DropdownMenuItem>
-
-          {/* Remove Audio */}
-          <DropdownMenuItem
-            data-testid="remove-audio-button"
-            onClick={handleRemoveAudio}
-            disabled={!audioFileName}
-          >
-            <Trash2Icon className="size-4" />
-            Remove Audio
-          </DropdownMenuItem>
-
-          {/* Export MIDI */}
-          <DropdownMenuItem
-            data-testid="export-midi-button"
-            onClick={handleExportMidi}
-            disabled={notes.length === 0}
-          >
-            <DownloadIcon className="size-4" />
-            Export MIDI
-          </DropdownMenuItem>
-
-          {/* Export ABC - File */}
-          <DropdownMenuItem
-            data-testid="export-abc-button"
-            onClick={handleExportABC}
-            disabled={notes.length === 0}
-          >
-            <DownloadIcon className="size-4" />
-            Export ABC
-          </DropdownMenuItem>
-
-          {/* Export ABC - Clipboard */}
-          <DropdownMenuItem
-            data-testid="copy-abc-button"
-            onClick={handleCopyABCToClipboard}
-            disabled={notes.length === 0}
-          >
-            <ClipboardIcon className="size-4" />
-            Copy ABC to Clipboard
-          </DropdownMenuItem>
-
-          <DropdownMenuSeparator />
-
-          {/* Auto-scroll toggle */}
-          <DropdownMenuCheckboxItem
-            data-testid="auto-scroll-toggle"
-            checked={autoScrollEnabled}
-            onCheckedChange={setAutoScrollEnabled}
-            onSelect={(e) => e.preventDefault()}
-            aria-pressed={autoScrollEnabled}
-          >
-            Auto-scroll
-            <DropdownMenuShortcut>Ctrl+F</DropdownMenuShortcut>
-          </DropdownMenuCheckboxItem>
-
-          {/* Debug toggle */}
-          <DropdownMenuCheckboxItem
-            data-testid="debug-toggle"
-            checked={showDebug}
-            onCheckedChange={setShowDebug}
-            onSelect={(e) => e.preventDefault()}
-          >
-            Debug
-          </DropdownMenuCheckboxItem>
-        </DropdownMenuContent>
-      </DropdownMenu>
+      {/* Settings button */}
+      <Button
+        data-testid="settings-button"
+        onClick={onSettingsClick}
+        variant="ghost"
+        size="icon"
+        title="Settings"
+      >
+        <SettingsIcon className="size-5" />
+      </Button>
 
       {/* Mixer button */}
       <Button
